requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
httpx>=0.27.0,<1
openai>=1.55.0,<2
pydantic-settings>=2.6.0,<3
python-dotenv>=1.0.0,<2
beautifulsoup4>=4.12.0,<5
tiktoken>=0.8.0,<1
cryptography>=43.0.0,<44
tzdata>=2024.1; sys_platform == "win32"

# tests
pytest>=8.0.0,<9
pytest-asyncio>=0.24.0,<1
